Moving f_text_block function to general io utilities.
